Easy Install For Building From Source
This is a pip freeze for all or most of the requirements that I could find when attempting to run the program. If there are more, feel free to add. Best way to determine if any are missing is to create a fresh vm and just run and pip install until there are no more complications.

How to use:

pip install -r requirements.txt

Standard practice in most python projects.