Verify that min price is not greater than max
Output error message and exit if reserve minimum price is greater than
reserve maximum price. Otherwise, the script will never actually
purchase anything because that's how math works.

Print price range with message that there were no offers that met it as
a further sanity check and to prevent ambiguity as to why the script
detected an offer but did not attempt to cart it